% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{addinexamplesRG}
\alias{addinexamplesRG}
\alias{addinexamplesRG-package}
\title{addinexamplesRG}
\description{
Supplies insertDotPipeAdin.
}
\details{
This package (\url{https://github.com/graiffgarcia/addinexamplesRG}) is a fork of \url{https://github.com/WinVector/addinexamplesWV}.
}
