% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinexamples-package.R
\docType{package}
\name{addinexamplesWV}
\alias{addinexamplesWV}
\alias{addinexamplesWV-package}
\title{addinexamplesWV (Win-Vector fork of \url{https://github.com/rstudio/addinexamples})}
\description{
Supplies insertDotPipeAdin.
}
