% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertBizzaroPipeAddin.R
\name{insertBizzaroPipeAddin}
\alias{insertBizzaroPipeAddin}
\title{Insert ->.; .}
\usage{
insertBizzaroPipeAddin()
}
\description{
Call this function as an addin to insert \code{->.;} (Bizarro Pipe, \url{http://www.win-vector.com/blog/2016/12/magrittrs-doppelganger/}) at the cursor position.
}
