% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertColonEqual.R
\name{insertColonEqual}
\alias{insertColonEqual}
\title{Insert := .}
\usage{
insertColonEqual()
}
\description{
Call this function as an addin to insert \code{:=} (Named map builder, \url{https://winvector.github.io/wrapr/reference/named_map_builder.html}) at the cursor position.
}
