% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertDotPipeAddin.R
\name{insertDotPipeAddin}
\alias{insertDotPipeAddin}
\title{Insert \%.>\% .}
\usage{
insertDotPipeAddin()
}
\description{
Call this function as an addin to insert \code{\link[wrapr]{\%.>\%}} (dot pipe, \url{https://winvector.github.io/wrapr/articles/dot_pipe.html}) at the cursor position.
}
