% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMagrittrPipeExpositionAddin.R
\name{insertExposition}
\alias{insertExposition}
\title{Insert exposition pipe operator}
\usage{
insertExposition()
}
\description{
Call this function as an addin to insert \code{\%$\%} (exposition pipe operator).
}
\seealso{
\code{\link[magrittr]{\%$\%}}.
}
