% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertLineOfOctothorpes.R
\name{insertLineOfOctothorpes}
\alias{insertLineOfOctothorpes}
\title{Insert forty octothorpes.}
\usage{
insertLineOfOctothorpes()
}
\description{
Call this function as an addin to insert 40 octothorpes, in order to create a
section break.
}
