% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMagrittrCompoundPipeAddin.R
\name{insertMagrittrCompoundPipeAddin}
\alias{insertMagrittrCompoundPipeAddin}
\title{Insert \%<>\% .}
\usage{
insertMagrittrCompoundPipeAddin()
}
\description{
Call this function as an addin to insert \code{\link[magrittr]{\%<>\%}} (Magrittr pipe, \url{https://github.com/tidyverse/magrittr}) at the cursor position.
}
