% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMagrittrEagerPipeAddin.R
\name{insertMagrittrEagerPipeAddin}
\alias{insertMagrittrEagerPipeAddin}
\title{Insert \%!>\% .}
\usage{
insertMagrittrEagerPipeAddin()
}
\description{
Call this function as an addin to insert \code{\link[magrittr]{\%!>\%}} (Magrittr pipe, \url{https://github.com/tidyverse/magrittr}) at the cursor position.
}
