% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMagrittrPipeAddin.R
\name{insertMagrittrPipeAddin}
\alias{insertMagrittrPipeAddin}
\title{Insert \%>\% .}
\usage{
insertMagrittrPipeAddin()
}
\description{
Call this function as an addin to insert \code{\link[magrittr]{\%>\%}} (Magrittr pipe, \url{https://CRAN.R-project.org/package=magrittr}) at the cursor position.
}
