% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertMagrittrTeePipeAddin.R
\name{insertMagrittrTeePipeAddin}
\alias{insertMagrittrTeePipeAddin}
\title{Insert \%T>\% .}
\usage{
insertMagrittrTeePipeAddin()
}
\description{
Call this function as an addin to insert \code{\link[magrittr]{\%<>\%}} (Magrittr pipe, \url{https://github.com/tidyverse/magrittr}) at the cursor position.
}
