% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsertPipeRPipeAddin.R
\name{insertPipeRPipeAddin}
\alias{insertPipeRPipeAddin}
\title{Insert \%>>\% .}
\usage{
insertPipeRPipeAddin()
}
\description{
Call this function as an addin to insert \code{\link[pipeR]{\%>>\%}} (default pipeR pipe, \url{https://CRAN.R-project.org/package=pipeR}) at the cursor position.
}
