% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertToDotPipeAddin.R
\name{insertToDotPipeAddin}
\alias{insertToDotPipeAddin}
\title{Insert \%>.\% .}
\usage{
insertToDotPipeAddin()
}
\description{
Call this function as an addin to insert \code{\link[wrapr]{\%>.\%}} (dot pipe, \url{https://winvector.github.io/wrapr/reference/grapes-greater-than-.-grapes.html}) at the cursor position.
}
