% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertUsrAddin.R
\name{insertUsr1Addin}
\alias{insertUsr1Addin}
\title{Insert user chosen text or operation.}
\usage{
insertUsr1Addin()
}
\description{
Call this function as an addin to insert user chosen text or operation at the cursor position.
To set the option do something like the example.
}
\examples{

# options("addinexamplesWV.usrFn1" = function() { rstudioapi::insertText(" := ") })

}
