% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertZeallotArrow.R
\name{insertZeallotArrow}
\alias{insertZeallotArrow}
\title{Insert \%<-\% .}
\usage{
insertZeallotArrow()
}
\description{
Call this function as an addin to insert \code{\%<-\%} (Zeallot Arrow, \url{https://CRAN.R-project.org/package=zeallot}) at the cursor position.
}
