% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_doc_with_preamble.R
\name{new_doc_with_preamble}
\alias{new_doc_with_preamble}
\title{Clean the environment and run the active script}
\usage{
new_doc_with_preamble()
}
\description{
Clean the environment and run the active script
}
