% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinexamples-package.R
\docType{package}
\name{addinexamplesWV}
\alias{addinexamplesWV}
\alias{addinexamplesWV-package}
\title{RStudio Addins (Win-Vector fork of \url{https://github.com/rstudio/addinexamples})}
\description{
Use RStudio addins to execute \R functions interactively within RStudio.
See \code{\link{findAndReplaceAddin}()} and \code{\link{insertInAddin}()}
for two examples of R functions that are exported.
}
