% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{addinexamplesWV}
\alias{addinexamplesWV}
\alias{addinexamplesWV-package}
\title{addinexamplesWV}
\description{
Supplies insertDotPipeAdin.
}
\details{
This package (\url{https://github.com/WinVector/addinexamplesWV}) is a fork of \url{https://github.com/rstudio/addinexamples}.
}
